requirements.txt
Django==5.0.2
python-dotenv==1.0.1
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
pika==1.3.2
drf-yasg==1.21.7
opencv-python==4.9.0.80
PyYAML==6.0.1
flower==2.0.1
docker==6.1.3
mlflow==2.15.1
psutil==6.0.7
python-socketio==5.11.3
daphne==4.1.2
uvicorn==0.30.6
channels
aiohttp==3.10.5